Backend/requirements.txt
boto3==1.21.46
Django==4.0.3
django-cors-headers==3.13.0
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-csv==2.1.1
dnspython==2.2.1
importlib-metadata==4.11.3
matplotlib==3.5.1
numpy==1.22.0
pandas==1.3.5
pyparsing==3.0.6
requests==2.27.1
s3transfer==0.5.2
s3utils==0.6.1
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
waitress==2.1.1
sqlparse==0.4.2